# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz API",
        "description": "API Documentation",
        "contact": {
            "name": "API Support",
            "email": "marko.mandic.engr@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/products/{product_id}/variants": {
            "get": {
                "tags": [
                    "Variants"
                ],
                "summary": "Get a paginated list of variants for a product",
                "description": "Returns a paginated list of variants belonging to a specific product",
                "operationId": "listProductVariants",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Variant"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Variants"
                ],
                "summary": "Create a new variant",
                "description": "Creates a new variant for a specific product",
                "operationId": "storeVariant",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID of the product",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Variant data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreVariantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Variant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Variant created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Variant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "handle": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The handle field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/variants/{id}": {
            "get": {
                "tags": [
                    "Variants"
                ],
                "summary": "Get a specific variant",
                "description": "Returns a specific variant with its details",
                "operationId": "getVariant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the variant to retrieve",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Variant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Variant not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Variants"
                ],
                "summary": "Update a variant",
                "description": "Updates a variant and returns the updated variant",
                "operationId": "updateVariant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the variant to update",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Variant data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateVariantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Variant updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Variant updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Variant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Variant not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "price": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The price must be a positive number."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Variants"
                ],
                "summary": "Delete a variant",
                "description": "Deletes a variant and returns a success message",
                "operationId": "deleteVariant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the variant to delete",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Variant deleted successfully"
                    },
                    "404": {
                        "description": "Variant not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreVariantRequest": {
                "required": [
                    "handle",
                    "price"
                ],
                "properties": {
                    "handle": {
                        "description": "URL-friendly slug for the variant",
                        "type": "string",
                        "example": "ergonomic-chair-red"
                    },
                    "price": {
                        "description": "Variant price",
                        "type": "number",
                        "format": "decimal",
                        "example": 219.99
                    }
                },
                "type": "object"
            },
            "UpdateVariantRequest": {
                "properties": {
                    "handle": {
                        "description": "URL-friendly slug for the variant",
                        "type": "string",
                        "example": "ergonomic-chair-blue"
                    },
                    "price": {
                        "description": "Variant price",
                        "type": "number",
                        "format": "decimal",
                        "example": 229.99
                    }
                },
                "type": "object"
            },
...
            "Variant": {
                "required": [
                    "id",
                    "product_id",
                    "handle",
                    "price"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the variant",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440002"
                    },
                    "product_id": {
                        "description": "ID of the product this variant belongs to",
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "handle": {
                        "description": "URL-friendly slug for the variant",
                        "type": "string",
                        "example": "ergonomic-office-chair-red"
                    },
                    "price": {
                        "description": "Variant price",
                        "type": "number",
                        "format": "decimal",
                        "example": 219.99
                    },
                    "created_at": {
                        "description": "When the variant was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    },
                    "updated_at": {
                        "description": "When the variant was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-07T14:30:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}